Project working, trting to fix image in the navbar
